Struct google_youtubeanalytics1::BatchReport

pub struct BatchReport {
    pub kind: Option<String>,
    pub report_id: Option<String>,
    pub time_span: Option<BatchReportTimeSpan>,
    pub outputs: Option<Vec<BatchReportOutputs>>,
    pub time_updated: Option<String>,
    pub id: Option<String>,
}

Contains single batchReport resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list batch reports (none)

Fields
kind: Option<String>

This value specifies the type of data of this item. For batch report the kind property value is youtubeAnalytics#batchReport.

report_id: Option<String>

The ID of the the report definition.

time_span: Option<BatchReportTimeSpan>

Period included in the report. For reports containing all entities endTime is not set. Both startTime and endTime are inclusive.

outputs: Option<Vec<BatchReportOutputs>>

Report outputs.

time_updated: Option<String>

The time when the report was updated.

id: Option<String>

The ID that YouTube assigns and uses to uniquely identify the report.