Struct google_youtubeanalytics1::ReportMethods

pub struct ReportMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on report resources. It is not used directly, but through the YouTubeAnalytics hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_youtubeanalytics1 as youtubeanalytics1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use youtubeanalytics1::YouTubeAnalytics;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = YouTubeAnalytics::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `query(...)`
// to build up your call.
let rb = hub.reports();

Methods

impl<'a, C, A> ReportMethods<'a, C, A>

fn query(&self, ids: &str, start_date: &str, end_date: &str, metrics: &str) -> ReportQueryCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieve your YouTube Analytics reports.

Arguments

ids - Identifies the YouTube channel or content owner for which you are retrieving YouTube Analytics data. - To request data for a YouTube user, set the ids parameter value to channel==CHANNEL_ID, where CHANNEL_ID specifies the unique YouTube channel ID. - To request data for a YouTube CMS content owner, set the ids parameter value to contentOwner==OWNER_NAME, where OWNER_NAME is the CMS name of the content owner.

start-date - The start date for fetching YouTube Analytics data. The value should be in YYYY-MM-DD format.

end-date - The end date for fetching YouTube Analytics data. The value should be in YYYY-MM-DD format.

metrics - A comma-separated list of YouTube Analytics metrics, such as views or likes,dislikes. See the Available Reports document for a list of the reports that you can retrieve and the metrics available in each report, and see the Metrics document for definitions of those metrics.